littlelemon/requirements.txt
asgiref==3.7.2
blinker==1.6.2
certifi==2023.11.17
cffi==1.16.0
click==8.1.3
cryptography==42.0.5
distlib==0.3.8
Django==5.0
djangorestframework==3.15.1
filelock==3.13.1
Flask==2.3.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pipenv==2023.11.15
platformdirs==4.1.0
pycparser==2.21
PyMySQL==1.1.0
six==1.16.0
sqlparse==0.4.4
virtualenv==20.25.0
Werkzeug==2.3.4
